Reorganizing Files
- Removing unnecessary App.css
- Moving helper files to main folder for better access
