+++ kanban-client/src/App.js
@@ -1,15 +1,10 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-
 import Kanban from './Components/Kanban'
 
-function App() {
+export default function App() {
   return (
     <React.Fragment>
       <Kanban />
     </React.Fragment>
   );
 }
-
-export default App;
